+++ app/src/main/res/layout/time_dialog.xml
@@ -137,7 +137,7 @@
 
         <CheckBox
             android:id="@+id/chAutoEnable"
-            android:layout_width="wrap_content"
+            android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:text="@string/auto_enable_label" />
